improve: origin checks for analytics
