+++ src/endpoints/types.ts
@@ -1,8 +1,9 @@
 import { SafeParseSuccess, z } from "zod";
-import { Env } from "../";
+import { Env } from "../worker";
 
 export type Endpoint<Schema extends z.Schema, Params = z.infer<Schema>> = {
   path: `/${string}`;
+  origins?: string[];
   schema: Schema;
   ship?: (
     params: SafeParseSuccess<Params>,
